requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pyyaml>=6.0
